requirements.txt
# LangChain core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic
anthropic

# HuggingFace Integration
langchain-huggingface
transformers
huggingface-hub

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# environment variable management
python-dotenv

# Machine learning utilities
numpy
scikit-learn
pandas

# Additional utilities
email-validator  # for Pydantic email validation
grandalf  # for LangChain graph visualization
langchain-groq